pyproject.toml
[tool.poetry]
name = "astra-vector-demo"
version = "0.1.0"
description = ""
authors = ["Yuki Morishita <mor.yuki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
langchain = "^0.0.279"
cassandra-driver = "^3.28.0"
cohere = "^4.23"
cassio = "^0.1.1"
streamlit = "^1.26.0"
python-dotenv = "^1.0.0"
openai = "^0.28.0"
tiktoken = "^0.5.1"
datasets = "^2.14.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
